requirements.txt
# Runtime
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Agent & AI
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Vector stores
faiss-cpu>=1.9.0
pgvector>=0.3.0

# Document loaders & embeddings
pydantic>=2.0.0
python-dotenv>=1.0.0

# MCP (Model Context Protocol) - stdio + streamable-http
mcp[cli]>=1.26.0

# Google Workspace MCP (Python, for ingestion source workspace_mcp)
workspace-mcp>=1.10.0

# DB (Railway)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Document ingestion
unstructured[md]>=0.16.0
python-docx>=1.0.0
pypdf>=5.0.0

# Google Drive SDK + OAuth ingestion
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# HTTP client (ingestion + MCP client)
httpx>=0.27.0

# Templates
jinja2>=3.1.0
